+++ crates/koto/examples/hello_world.rs
@@ -3,7 +3,7 @@ use koto::{Result, prelude::*};
 fn main() -> Result<()> {
     let script = "print 'Hello, World!'";
 
-    Koto::default().compile_and_run(script)?;
+    Koto::with_settings(KotoSettings::default().inherit_io()).compile_and_run(script)?;
 
     Ok(())
 }